+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.4.0 (TBD)
+### Enhancements
+
+### Bug fixes
+
+
+## 0.3.0 (2020-12-10)
+
+### Enhancements
+  - Improved objectives page with new "break down" feature
+  - Add API documentation
+### Bug fixes
+  - Fix an LTI 1.3 issue where launch was using kid to lookup registration instead of issuer and client id
+
 ## 0.2.0 (2020-12-10)
 
 ### Enhancements